Made registry of `chkrec_keeper`s.
Provides a convenience function to spawn new `chkrec_keeper`s under
the supervision of the `chkrec_keeper_sup` process, while maintaining
a mapping of the identifiers for useful workers to the pids of the
backup keepers. A useful worker can employ this registry to lookup the
pid for the backup keeper and store/retrieve its state.

Note that if for any reason the `chkrec_keeper` were to go down, the
information stored in registry would be stale. This will be fixed in
a later commit.