Added `chkrec_keeper_mon:watch/2` function.
Calls `erlang:monitor/2` to keep track of when a `chkrec_keeper`
process goes down. This is done so that subsequent calls to
`chkrec_keeper_mon:lookup/1` will return the new pid. However, the
calls to `chkrec_keeper_mon:watch/2` and `chkrec_keeper_mon:update/2`
still need to be integrated into the initialization phase of the
backup keeper (i.e. `chkrec_keeper:init/1`) so the new pid is
actually stored when the process is restarted.

Note that the entry to the backup keeper is never removed, but can
only be overwritten. A better approach may be to only change the state
to `down` if the process had exited abnormally, and remove the entry
otherwise. (We can do this since we are using a `transient` restart
strategy for the `chkrec_keeper`s.)