Integrated monitoring into `chkrec_keeper_mon:update/2`.
Now if the `chkrec_keeper` process changes for a particular source,
then the new process is monitored. The normal occurrence for this
would be the transition `down ~~> <pid>` after the backup keeper is
restarted.

However, the case for when a backup keeper is currently running is
also handled by demonitoring that process, and then terminating it
with a "normal" reason, i.e. `{shutdown, for_update}`. Doing so is
a more expensive operation because the entire `dict` is traversed
through to find the key corresponding to a certain value.