Fixed a couple issues in the `chkrec_keeper_mon` module.
No need to call `chkrec_keeper_mon:update/2` within
`chkrec_keeper_mon:new/1` since `chkrec_keeper:init/1` will do so.
This would actually cause a second restart to occur.

Was storing the pid itself, not {value, <pid>} in the sources registry,
so the case to demonitor and terminate the old process was not being
executed.